+++ src/main/java/com/example/sweater/repos/MessageRepo.java
@@ -1,8 +1,3 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by Fernflower decompiler)
-//
-
 package com.example.sweater.repos;
 
 import com.example.sweater.domain.Message;
@@ -10,7 +5,7 @@ import com.example.sweater.domain.Topic;
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTag(String tag);
 
     List<Message> findMessageByBelongsToTopic(Topic topic);
